requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
asyncpg>=0.29.0
aiosqlite>=0.19.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
slowapi==0.1.9
google-generativeai==0.8.5
sentence-transformers==5.1.0
faiss-cpu==1.12.0
numpy==2.3.2
email-validator>=2.0.0
pytest==8.4.2
httpx==0.28.1
requests

# Speech & Audio Processing
openai-whisper
pyannote.audio
torch
torchaudio


# Database Migrations
alembic==1.13.1

# Background Job Processing
celery[redis]==5.3.6
redis==5.0.1

# Monitoring & Metrics
prometheus-client==0.20.0

# Testing
pytest-asyncio==0.23.5
pytest-cov==4.1.0
